Break up ModelGeneratorType.scala
