+++ modules/scala-dropwizard/src/main/scala/dev/guardrail/generators/scala/dropwizard/DropwizardVersion.scala
@@ -1,0 +1,9 @@
+package dev.guardrail.generators.scala.dropwizard
+
+sealed abstract class DropwizardVersion(val value: String)
+object DropwizardVersion extends DropwizardVersion("dropwizard") {
+  def unapply(version: String): Option[DropwizardVersion] = version match {
+    case "dropwizard" => Some(DropwizardVersion)
+    case _            => None
+  }
+}
