+++ modules/scala-support/src/main/scala/dev/guardrail/generators/scala/JacksonModelGenerator.scala
@@ -1,0 +1,9 @@
+package dev.guardrail.generators.scala
+
+sealed abstract class JacksonModelGenerator(val value: String) extends ModelGeneratorType
+case object JacksonModelGenerator extends JacksonModelGenerator("jackson") {
+  def unapply(version: String): Option[JacksonModelGenerator] = version match {
+    case JacksonModelGenerator.value => Some(JacksonModelGenerator)
+    case _                           => None
+  }
+}
